requirements.txt
langchain==0.2.6
numpy==1.26.4
torch==2.3.1
transformers==4.42.3
sentence-transformers==3.0.1
beautifulsoup4==4.12.3
langchain-community==0.2.6
langchain-groq==0.1.6
streamlit==1.22.0
python-dotenv==1.0.0
faiss-cpu